requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
rembg[cpu]>=2.0.50
pillow>=10.2.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
slowapi>=0.1.9
